==> 1 - Data Cleaning - NFL.Rmd <==
# Data Cleaning and Formatting
##  Juan Solorio
##  Data 598C Wi2020

I will be using data obtained from 'http://nflsavant.com/about.php', which provided data from 2013 to 2019 about the 'Play-by-Play' description of each NFL game from Week 1 to the Super Bowl. 

Let's load each file into and check the dimesions of the files.
```{r}
#   Setting the working directory where files are stored
setwd('C:/Users/Juan Solorio/Desktop/DataScience/Grad School/Classes/Data 598C/')

#   Loading the files into individual data frames
S2013 <- read.csv('pbp-2013.csv')
S2014 <- read.csv('pbp-2014.csv')
S2015 <- read.csv('pbp-2015.csv')
S2016 <- read.csv('pbp-2016.csv')
S2017 <- read.csv('pbp-2017.csv')
S2018 <- read.csv('pbp-2018.csv')
S2019 <- read.csv('pbp-2019.csv')

#   Creating a 'dictionay' so I can iterate through the dfs and see the dimensions and structure of each
Seasons_dict <- list(S2013, S2014, S2015, S2016, S2017, S2018, S2019)

#   Displaying the Dimensions and structure of each file

for (season in c(1:7)) {
  print(c('Season: ', 2012+season, 'Dim (col, row): ',dim(Seasons_dict[[season]])))
}

```
From the dimensions of the data frames, all have 45 columns and most are aroung 40k rows. The only exception is the 2013 season df, which only had about 15k rows, I think I will drop the season and keep the rest as it might allow for a better speed up of the rendering. Now lets check that all the columns match and are in the same order.

```{r}
# Checking the data frames have the same names for columns, a mean of 1 means they are the same
for (season in c(1:6)) {
  print(c('Season: ', 2012+season, 'Same columns?: ',ifelse(mean(names(Seasons_dict[[1]]) == names(Seasons_dict[[season]]))==1,yes = 'Yes', no = 'No')))
}

```
Since all the columns are the same for all the seasons, I will proceed to set my train and test set. For this I will combine all seasons from 2014 to 2018 as a training set and leave the 2019 for testing.

```{r}
# specifying the test and train sets
data_test <- S2019
data_train <- rbind(S2014, S2015, S2016, S2017, S2018)
```
Lets check each of the sets and view the data.

```{r}
# Calling out each df to have a glance at the actual data
View(data_test)
View(data_train)

# lets also check the number of NAs each column has, first I'll make sure that '(black)' rows are NAs
data_train[data_train == ''] <- NA
data_test[data_test == ''] <- NA
check_df_nulls <- function(df) {
  ## Function to check the number of NAs in a df
  for(i in 1:dim(df)[2])
  {
    cat(sprintf('Column %.0f: %30s \t Number of NAs: %.0f \t Percent NA data: %.0f%% \n', i,names(df)[i], length(which(is.na(df[,i]))),100*length(which(is.na(df[,i])))/length(df[,i])))
    
  }
}

# Checking the number of NAs in each df
check_df_nulls(data_test)
check_df_nulls(data_train)
```
From the data there seems to be a few columns without actual values or names labeled as 'X', 'X.1', 'X,2', 'X.3', 'Challenger'. There is also three columns with a large amount of data missing (>70%). I will drop these since they are not going to contribute to the actual project and it would be hard to figure out the actual data entry from the other columns. 
The column 'OffenseTeam' is one which we might be able to figure out the 7% of unknowns. Each game has a GameId and there are two teams, one on offense and one on defense per each row. I could filter the results for game and then fill in the Offense team missing by checking the team that is on defense for that game. 

```{r}
# columns to drop 'X', 'X.1', 'X,2', 'X.3', 'Challenger'
data_train <- data_train[, -c(11, 13, 17, 18, 31)]
data_test <- data_test[, -c(11, 13, 17, 18, 31)]

# Using dplyr library to 
library(dplyr)
    ## Getting the ids for games
gameids_test <- unique(data_test$GameId)
gameids_train <- unique(data_train$GameId)

teams_in_game <- function(df, column1, defense, offense, gameids) {
  # Function to replace the NAs of OffenseTeams for the offense team opposite to Defense team in a game
  for (gameid in gameids) {
    #print(gameid)
    teams <- unique(df[df[column1] == gameid,][,defense])
    #print(teams)
    tempdf <- df[(df[column1] == gameid & is.na(df[offense])),]
    #print(nrow(tempdf[column1]))
    for (i in 1:nrow(df[(df[column1] == gameid & is.na(df[offense])),][column1])) {
      ifelse(df[(df[column1] == gameid & is.na(df[offense])),][i,defense] == teams[1],df[(df[column1] == gameid & is.na(df[offense])),][i,offense] <- teams[2],df[(df[column1] == gameid & is.na(df[offense])),][i,offense] <- teams[1])
     #df[(df[column1] == gameid & is.na(df[offense])),] <- tempdf
    }
    
  }
  return(df)
  
}

data_test <- teams_in_game(data_test,column1 = 'GameId', defense = 'DefenseTeam', offense = 'OffenseTeam',gameids = gameids_test)

data_train <- teams_in_game(data_train,column1 = 'GameId', defense = 'DefenseTeam', offense = 'OffenseTeam',gameids = gameids_test)

```
Now lets look at the data type of each column for the data frames, might need to adjust depending on the
data that each column contains.

```{r}
# Checking the structire of the tables, finding out what each data type is
str(data_train) 
str(data_test)
```
Both tables have a mix between factors and ints as the data types for the columns. Factors would be useful 
when trying to manipulate categorical data. However I think I will need to change columns such as 'Description'
to a character since it is text data and 'GameDate' to date format. I will also need to change some of the int data columns to factors such as the 'Quarter', 'Minute', 'Down', 'SeriesFirstDown', and all columns after 'Formation' as they act as boolean describing if something was accomplished in the play.

```{r}
# I will first get he column number for each of the columns to convert to a factor, and make to a list
      # I've gather that the columns to remain as factors are by number
      # 1 , 2, 3, 6, 7, 8, 11, 13, 17 : 40
factor_columns <- c(1 , 2, 3, 6, 7, 8, 11, 13, 17 : 40)

data_test[factor_columns] <- lapply(data_test[factor_columns], factor)
data_train[factor_columns] <- lapply(data_train[factor_columns], factor)

# Checking it all worked out
str(data_train) 
str(data_test)

# Now to convert the 'Description' from factor to character and Gamedate to time
data_test$Description <- as.character(data_test$Description)
data_train$Description <- as.character(data_train$Description)
data_test$GameDate <- as.Date(data_test$GameDate, format = '%Y-%m-%d')
data_train$GameDate <- as.Date(data_train$GameDate, format = '%Y-%m-%d')


# Checking it all works
str(data_train) 
str(data_test)
```
The columns for now are in the format we want for doing some manipulation, for now lets just sort them by 
'GameId', 'Quarter','Minute', 'Second' to have a better grouping and sense of passage of time in the game.

```{r}
# Ordering the tables by Ascending 'GameId', 'Quarter' and decensing 'Minute', 'Second' 
data_test <- data_test[order(data_test$GameId, data_test$Quarter, -data_test$Minute, -data_test$Second),]
data_train <- data_train[order(data_train$GameId, data_train$Quarter, -data_train$Minute, -data_train$Second),]
```
Lets now take a glimpse at the summary for each of the columns.

```{r}
summary(data_train)
summary(data_test)
```
We still have plenty of NAs in columns like 'OffenciveTeam', 'Formation', 'PlayType', 'PassType', 'RushDirection', 'PenaltyTeam', and 'PenaltyType'. For now lets take a look at some graphs to have visual look at the distribution and values for the various columns. 
I will just run a simple bar plot to view the number of events for column per value.

```{r}
######################################
#           Saving Data
######################################

save(dat_train, file = "data_train_cleaned.RData")

###########################
#     "Histograms"

for (col in 2:ncol(data_test)) {
    barplot(table(data_test[,col]), main = c('Plays by :',names(data_test)[col]))
}

```
